requirements.txt
streamlit==1.52.2
tensorflow==2.20.0
keras==3.13.0
numpy==2.4.0
Pillow==12.0.0
